pyproject.toml
[tool.poetry]
name = "mydownloaderbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.1"
replit = "^2.0.0"
mutagen = "^1.45.1"
validators = "^0.18.2"
wget = "^3.2"
pytube = "^12.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
